requirements.txt
chromadb==1.3.0
discord.py==2.6.4
duckduckgo_search==8.1.1
langchain_core==1.0.2
langchain_groq==1.0.0
openai==2.6.1
pytest==8.4.2
python-dotenv==1.2.1
Requests==2.32.5